O622-002 Fix of broken test after the commit for counter examples.
The test failed because gnatwhy3 always loaded a prover for a
counter-example query (cvc4_ce) and the entry for this prover was not
generated in test.whyconf. And in fact, it makes no sence to generate it
there in this case.

This commit solves this issue by loading the counter-example prover only
if counter-example will be queried (--couter-example option is passed
to gnatwhy3).

* src/gnat/gnat_config.ml
Loading counter-example prover only if an option --counter-example=on is
passed to gnatwhy3