N318-011 Filtering out superfluous locations from counter-example.
As discussed in O626-003, if gnatprove is run with an option
--proof=per_path, there there can be elements in counter-example that
are not in the trace corresponding to the path. These elements have
values that make the counter-example obscured.

This commit removes these elements from generated counter-example.

* src/core/model_parser.ml
(model_for_positions): Given a model and a list of source-code positions
returns a model with elements on these positions.
* src/gnat/gnat_main.ml
(handle_vc_result): Passing of trace to the point where a
counter-example is saved
* src/gnat/gnat_objectivesml
(save_counter_example): Filtering the model before it is saved.