N318-011 O626-003 Bugfix.
If a trace contains one element with a single location for the goal, other
locations should not be filtered out from the counter-example.

* src/gnat/gnat_objective.ml
(save_trace): Return Gnat_loc.S.empty if no tracefile was generated.
(save_counterexample): Do not filter locations that are not in trace if
trace is Gnat_loc.S.empty.

* src/gnat/gnat_objective.mli
(save_trace): Documentation changed to correspond the implementation.