N318-011 Fix to losing locations in fast_wp (implemented by Johannes).
The location tags in terms corresponding to constructs that trigger VC were
lost in some cases in the process of fast wp. This caused not displaying
information in the lines of these terms in gps.

* src/whyml/mlw_wp.ml
(wp_label): added boolean parameter override. If it is true, the location
of the term is overriden by the location of expression inside that the
term is.
(fast_wp_desc): in case that expression is Eapp, the location of the pre
term is overriden by the location of the expression.