O827-013 Do not display prover errors when getting counterexample.
When cvc4 is run to get counterexample, it can fail and return error. For
example, to get counterexample, the logic needs to be set to AUFBVLIRA
(current limitation of cvc4) and when there is some non-linear fact,
cvc4 fails with error message.

This commit makes gnatprove to not display these errors.

* src/gnat/gnat_main.ml
(interpret_result): add gnatprove warning with prover answer only if
getting counterexample is off