N318-011 O619-018 Adding attribute 'Old to field accesses.
In the case of field access, the attribute 'Old is added after the name
of the enclosing variable, not after the name of the field.

* src/gnat/gnat_objectives.ml
(spark_counterexample_transform): In the case of field access, add the
attribute 'Old after the name of the enclosing variable, not after the
name of the field.