N318-011 Store counterexample information related to VC in special index.
Since in gnatwhy3, the line number of the construct that triggers VC is
not known, map the counterexample information related to this construct
to index "vc_line" instead of mapping it to line number.

Note that in gnatwhy3, the line of the construct that triggers VC is
guaranteed to be known only if this construct does not span to multiple
lines or if the VC is not split.

src/gnat/gnat_report.ml
(print_cntexmp_model): Set the index where the counterexample information
about the conctruct that triggers VC are stored to "vc_line".