N318-011 Do not include record fields without traceability information.
Record fields without traceability information are now not included in
counterexamples. These are fields without "model_trace:*" label.

Note that if a record field has label "model_trace:", the name of the
record field will be skipped (but the record field will be included in the
counterexample). This is usually used when the actual content of the
counterexample element is stored in a record internal to why3
representation of SPARK program. If the record field has label
"model_trace:@hide_field", the field will not be included in the
counterexample at all.
Finally, in Inria why3, if the record field has no
"model_trace:*" label, it will be displayed in counterexample and why3
name of the field will be used. On the contrary, gnatwhy3 does not include
the field in counterexample as if the label would be
"model_trace:@hide_field". The rationale behind this is that only record
fields that have explicit traceability information should be included in
counterexample.
Note that there is a number of record fields defined in
"spark2014/share/spark/theories" that are not yet supported by
counterexample generation and that should be thus not included in
counterexamples.