OA21-017 Fix of failure in_of_string on Windows.
The file of the form "file:line:col1:col2" was created, than split on ":"
and and int_of_string was used to get line number and colon numbers. This
failed on windows because the file name contained ':', i.e.,
"C:\...\gnatprove\p\..\p.mlw:2254:55:150".

This shouldn't happen, because we always use short file names (no full
file names). So either not enough why locations are generated by gnat2why
(in this case the terms have locations from generated *.mlw file, not from
original SPARK code) or why location is incorrectly propagated to the term
in WP. To be investigated.

This fix is authored by Johannes.