Displaying of counterexamples enhanced.
When counterexample is displayed interleaved with source code, only the
part of the source code from the beginning to the line of the last
counterexample element is displayed and the focus in the editor is on the
end of that source code.
Textual counterexample (without interleaving with source code) is not
displayed by default. There is a debug flag debug_show_text_cntexmp that
enables such display.