Collect the last model element with given name and location.
Counterexample model elements with the same name and location should
not be displayed together. Only the one that corresponds to the term
that is later in the task should be displayed. There can be two
counterexample elements with the same name and location if why  code
is generated from some source language and location are locations
in the source language. This happens e.g., if why code is generated
from SPARK. There, the first iteration of while cycle is unrolled in
some cases. If the task contains both a term representing a variable
in the first iteration of unrolled loop and a term representing the
variable in the subsequent loop iterations, only the latter is
relevant for the counterexample and it is the one that comes after
the former one.