OB03-034 OA19-023 Collect the last element with given name and location.
Collect the last counterexample element with given name and location in
why3.

The commit 6daedc73793946d5b3df52f53706335b0ed8694f solved the problem
that more counterexample elements with the same name and location had
different labels. This can happen either if some of them should be
projected and some of them not (label "model" vs label "model_projected"
of if they were created in different type of expression in WP). In these
cases, these counterexample elements were present in a counterexample
sent back to gnat2why. However, if counterexample elements have the same
labels and the same location, only the first one was collected by why3,
which was not correct (only the last one should be collected).

Counterexample model elements with the same name and location should
not be displayed together. Only the one that corresponds to the term
that is later in the task should be displayed. There can be two
counterexample elements with the same name and location if why  code
is generated from some source language and location are locations
in the source language. This happens e.g., if why code is generated
from SPARK. There, the first iteration of while cycle is unrolled in
some cases. If the task contains both a term representing a variable
in the first iteration of unrolled loop and a term representing the
variable in the subsequent loop iterations, only the latter is
relevant for the counterexample and it is the one that comes after
the former one.

* src/printer/smtv2.ml
While collecting terms corresponding to counterexample elements, collect
the last term with given location and and counterexample name.