OB10-023 Enable transformation detect_polymorphism for counterexamples.
For variables of boolean type, sometimes, just references to internal
structures were get as counterexample values. This was due to the encoding
of why boolean type to smt. When the transformation detect_polymorphism is
used, the encoding of boolean type is always suitable for getting
counterexamples.

* drivers/cvc4_15.drv
Enable the transformation detect_polymorphism and change the logic to
support datatypes, which is necessary for encodings that are used when
detect_polymorphism is enabled.