OC14-006 do not store counter example run results in session
Why3 relies on session information, so the session file should always be
in the correct state. Here however we were storing the result of counter
example run in the session, but later ignoring it for proof results.
This resulted in bogus output.

* gnat_sched.ml
(handle_finshed_call): set results of CE prover run to unknown