cpulimit-win: produce the same output as cpulimit.c
Drivers that parse the cpulimit output expect a line of the form
"why3cpulimit time : XXX s" for the total cpu time of the prover run.

The Win version of cpulimit prints an additional line
"why3cpulimit real time : XXX s" for the wall-clock time.
This line is ignored by drivers, since we use the cpu time
and not the real time to establish time limits.