OA20-004 fix duplicate message with counter examples
When option -j > 1 was given, gnatwhy3 would run several provers on
sub-VCs of same check, and report this check as unproved several times (if
some or all of the proofs are unsuccesful). This is mostly not a problem
because identical messages are suppressed by the GNAT messages
machinery.

However, gnatwhy3 would also run counter-examples for each of the failed
proof attempts above, which not only wastes CPU time, but also makes
each of the failed messages different, resulting in duplicate messages.

This commit now does:
1) do only one counter-example prover run per check;
2) only provide a single message per check to gnat2why

* gnat_objectives.ml
(register): do not reset counter_examples variable to false, so that
  acounter examples are only run once per check
*gnat_report.ml)
(register_result): the list of messages becomes a map from checks to
  messages; discard result when check is already in map
(print_json_msg): adapt to data structure change