P219-012 allow gnatwhy3 run without provers
The objective of this commit is to make gnatwhy3 work with an empty
prover list (and still report messages). The main problem is that
previously, all reporting would be done only when a prover returned,
because it was the place where we had all the information conveniently
at hand. So not running any prover would result in no messages.

In this commit, we change the way gnatwhy3 works and do the reporting at
the end, regardless of prover runs. We use the session to find all the
information we need (goal proved/not proved, statistics, find the right
task/proof attempt to extract extra info), and pass it on to
gnat_report.

* gnat_config.ml
(prover configuration): no default prover list in gnatwhy3 anymore;
allow empty list of provers now
* gnat_main.ml
(handle_vc_result): vastly simplified, doesn't do any reporting now
(report_messages): new function to report the message for a given
objective. Does all the computation that previously was done in
handle_vc_result
(main): call report_messages on all objectives
*gnat_objectives.ml
(session_proved_status): new function to determine proved/not proved
status of an objective only by looking at the session
(session_find_unproved_pa): new function to find a suitable unsuccessful
proof attempt to extract the right information
*gnat_report.ml
(register): almost all arguments of this function were only used in one
of the cases (proved/not proved). So replaced the long argument list by
a variant type