P715-001 read why3.conf from static location
Instead of reading why3.conf from current dir (gnatprove subdir in
fact), read it from a static location in the SPARK install.

This new why3.conf file doesn't contain any loadpath any more, so need
to add those loadpath manually. They are static, though.

The other problem are driver locations, but in fact thery are either
static (pre-installed drivers) or relative to the config file that
mentions them, so it's not a problem either.

* gnat_config.ml
  - add paths that were previously part of the why3.conf file to the
    loadpath programmatically;
  - Rewrite driver paths returned by the why3 api to point to correct
    driver location in the case of pre-installed provers