Dexpr: do not check for redundant annotations under "let rec"
Dexpr.check_spec may erroneously reject a valid user annotation
for a side effect of a local function that recursively calls its
parent function, since the full effect of the parent function is
not yet known when check_spec is called. Therefore we omit this
check for local functions defined inside recursive functions.