P720-007 simplify SPARK specific prover record
This record type is almost identical to an existing type in the Why3
API, save for the 'editor' field. This field was used in only one place,
so we can query for the value at that point.

* gnat_config.ml
simplify prover type
(init): remove editor field on record init, and the code which was needed
  for this init
* gnat_manual.ml
(editor_command): code from gnat_config.init moved and inlined here

Change-Id: I070cd9cfdeca8f55b46fab6242f78833d2762f89