P915-042 fix stack overflow in gnatwhy3
The collect_cases function was written in way that could lead to
concatenation of large lists, which itself could lead to stack overflow.

Deceptively, collect_cases was written in the "accumulator" style, which
is usually applied to avoid this problem, but here the technique was not
well applied and didn't help.

This commit rewrites the collect_cases using "real" accumulator style,
which now avoids a stack overflow.

* gnat_split_disj.ml
(collect_cases): rewrite

Change-Id: I2b3df6bf63e594d925cdac81ef50ec218fa7c220