Adding a search command.
Added a last set which is build at the same time of the name_tables which
is a map from id to list of decls where id is used. This has to be tested
if this is actually efficient (time and memory). If not, we can still
disallow it with a boolean.