PB18-021 fix replay with custom why3.conf file
If replay is used together with option --why3-conf, we need to read that
file either when an external prover has been explicitly provided, or
when the --prover option has not been provided at all. The latter case
did not work, fixing this here.

* gnat_config.ml
(top_level): force reading extra config file when replay is presentt

Change-Id: If2cb254926e6932df7ba993a646532fc817a0c41