PC05-004 disable record expansion for large records
the add_quant function is super slow on large records, and the eval
function which is called on the result of add_quant, is even slower.

We disable the add_quant function if it would expand a record of more
than 20 fields.

* eval_match.ml
(count_record_fields): function to count fields of a record type
(add_quant_small): wrapper for add_quant that does nothing if
count_record_fields is larger than threshold.

Change-Id: Iebab4962432f4dd65eb0e99d3948b353bb02b21d